Distinguish null result from error
When a remote endpoint replies with e.g. `{"jsonrpc":"2.0","result":null,"id":1}` we get a `None` from `next_repsponse()`. This is not an error so the current implementation is a bit confusing.
`common::Response` (and Serde) doesn't parse `null` as a valid JSON document (see https://stackoverflow.com/questions/8526995/is-null-valid-json-4-bytes-nothing-else) so this PR replaces `null` with `[]` which does parse. Not elegant but perhaps better than the status quo?