refactor(common): rename common -> types..
This commit renames the `common module` to `types` and tries to distinguish the types that is
directly related to the `JSON-RPC v2 specification` from others.

Somethings are a little big sloppy named as naming is hard.

Also, as bonus a removed a bunch of needless stuff in http server related to subscription.