fix(websocket client): drop subscriptions that can't keep up with the internal buffer size (#166)
* fix(ws client): drop subscriptions when full.

This commit changes the behavior in the `WebSocket Client` where each subscription channel is used in a non-blocking
matter until it is determined as full or disconnected. When that occurs the channel is simply dropped and when
the user `poll` the subscription it will return all sent subscriptions before it was and terminate (return None)
once it's polled one last time. Similarly as `Streams` works in Rust.

It also adds configuration for the `WebSocket Client` to configure capacity for the different internal channels to avoid
filling the buffers when it's not expected.

* tests(ws client): simple subscription test.

* fix: nits

* Update src/client/ws/client.rs

* refactor(tests): introduce integration_tests

Make the repo structure more understable w.r.t testing.

* chore(license): add missing license headers

* Update src/client/ws/client.rs

* Update src/client/ws/client.rs

* style: remove unintended spaces.

* tests: add concurrent deadlock test

Ensure that if more than the requested channel buffer capacity is exceeded it should not deadlock.
Such as spawning alot of concurrent requests, notifications or new subscriptions.

* Update src/client/ws/client.rs

* fix: review grumbles

* fix nits: `remove needless closure`

* fix: cargo fmt

* Update src/client/ws/client.rs

Co-authored-by: David <dvdplm@gmail.com>

* fix more nits

Co-authored-by: David <dvdplm@gmail.com>