address grumbles: change to Result<(), Option<M>>
This changes the subscription API again to return `Result<(), Option<SubscriptionMessage>>`
to work a little smoother with the combinators and it introduces an extention trait
to make convert it `Result<T, Error>` to `Result<T, Option<SubscriptionMessage>>`

This trait is implemented for types where it is possible and for custom types and tricky situations the
user has to implement themselves.

For instance it's not possible to implement on `TrySendError` because that depends on the use-case
i.e. whether the channel is full should be regarded as an error or not.