client(subscription): improve stream impl
It's was hard to know whether a subscription was closed or that the
entire connection was closed without calling `Client::is_connected`.

This commit adds an explicit Error::Lagged to indicate that
the subscription was dropped because it could not keep with the server.

Hopefully, the explict error and added documentation will help
with any confusion regarding that.