fix(proc macros): don't use `params` as name
Because jsonrpsee's internal proc macro code was using `params` it could
conflict and with `params` in the user's API in such case it
may try to overwrite the "wrong" variable.

In this PR I have renamed this internal variable to something much
weirder which should be quite unlikely that anyone will use.