core: impl ToParams for json::Value
This PR is starting the work to allow a `{...}` as a Params.

The related part of the spec is https://www.jsonrpc.org/specification

```
4.2 Parameter Structures
If present, parameters for the rpc call MUST be provided as a Structured value. Either by-position through an Array or by-name through an Object.

by-position: params MUST be an Array, containing the values in the Server expected order.
by-name: params MUST be an Object, with member names that match the Server expected parameter names. The absence of expected names MAY result in an error being generated. The names MUST match exactly, including case, to the method's expected parameters.
```

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>