feat(server): add `Methods::inject_extensions`
The RpcModule/Methods may need to inject extensions which could be injected
by the implementation as Extensions.

This PR adds additional API to be used for unit testing a RpcModule.
For instance one may inject some arbitary data which can't be unit tested
unless it's explictly injected by the RpcModule.